# docs/openapi/phase-1.yaml
...
openapi: 3.1.0
info:
  title: UB — Phase 1 API
  description: Identity, tenancy, and catalog backbone (Phase 1).
  version: 0.0.1-SNAPSHOT
servers:
- url: /
  description: Current host
paths:
...
  /api/v1/items:
    get:
      tags:
      - items-controller
      operationId: list_1
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: barcode
        in: query
        required: false
        schema:
          type: string
      - name: categoryId
        in: query
        required: false
        schema:
          type: string
      - name: includeCategoryDescendants
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: noBarcode
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: includeInactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageItemSummaryResponse'
    post:
      tags:
      - items-controller
      operationId: create_2
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ItemResponse'
...
  /api/v1/items/{id}/variants:
    post:
      tags:
      - items-controller
      operationId: createVariant
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVariantRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ItemResponse'
  /api/v1/items/{id}/images:
    post:
      tags:
      - items-controller
      operationId: registerItemImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterItemImageRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ItemImageResponse'
  /api/v1/items/{id}/images/upload:
    post:
      tags:
      - items-controller
      operationId: uploadItemImageCloudinary
      summary: Upload a product image to Cloudinary (stores URL + metadata only in DB)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image bytes (JPEG, PNG, WebP, GIF, …)
                altText:
                  type: string
                  maxLength: 500
                primary:
                  type: boolean
                  default: true
                  description: When true, sets item cover imageKey to the new secure URL
              required:
              - file
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ItemImageResponse'
        '503':
          description: Cloudinary not configured on this server
...
  /api/v1/items/{id}/pricing-context:
    get:
      tags:
      - items-controller
      operationId: getItemPricingContext
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EffectivePricingContextResponse'
  /api/v1/items/{id}:
    get:
      tags:
      - items-controller
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ItemResponse'
    delete:
      tags:
      - items-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: cascadeVariants
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        '204':
          description: No Content
    patch:
      tags:
      - items-controller
      operationId: patch_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchItemRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ItemResponse'
...
  /api/v1/items/{id}/images/{imageId}:
    delete:
      tags:
      - items-controller
      operationId: deleteItemImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: imageId
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No Content
...
components:
  schemas:
...
    CreateItemRequest:
      type: object
      properties:
        sku:
          type: string
          maxLength: 191
          minLength: 0
        barcode:
          type: string
          maxLength: 191
          minLength: 0
        name:
          type: string
          maxLength: 500
          minLength: 0
        description:
          type: string
          maxLength: 10000
          minLength: 0
        itemTypeId:
          type: string
          maxLength: 36
          minLength: 0
        categoryId:
          type: string
          maxLength: 36
          minLength: 0
        aisleId:
          type: string
          maxLength: 36
          minLength: 0
        unitType:
          type: string
          maxLength: 16
          minLength: 0
        isWeighed:
          type: boolean
        isSellable:
          type: boolean
        isStocked:
          type: boolean
        packagingUnitName:
          type: string
          maxLength: 255
          minLength: 0
        packagingUnitQty:
          type: number
        bundleQty:
          type: integer
          format: int32
        bundlePrice:
          type: number
        bundleName:
          type: string
          maxLength: 255
          minLength: 0
        minStockLevel:
          type: number
        reorderLevel:
          type: number
        reorderQty:
          type: number
        expiresAfterDays:
          type: integer
          format: int32
        hasExpiry:
          type: boolean
        imageKey:
          type: string
          maxLength: 2048
          minLength: 0
      required:
      - itemTypeId
      - name
      - sku
    ItemImageResponse:
      type: object
      properties:
        id:
          type: string
        s3Key:
          type: string
          description: Legacy storage key; Cloudinary rows often duplicate publicId here
        secureUrl:
          type: string
          maxLength: 2048
        publicId:
          type: string
          maxLength: 512
        provider:
          type: string
          description: legacy | cloudinary
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        sortOrder:
          type: integer
          format: int32
        contentType:
          type: string
        altText:
          type: string
        bytes:
          type: integer
          format: int64
        format:
          type: string
        assetSignature:
          type: string
          description: Provider version / etag string
        predominantColorHex:
          type: string
          description: Dominant color from Cloudinary colors analysis
        phash:
          type: string
          description: Perceptual hash for deduplication workflows
        createdAt:
          type: string
          format: date-time
    ItemResponse:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        barcode:
          type: string
        name:
          type: string
        description:
          type: string
        variantOfItemId:
          type: string
        variantName:
          type: string
        categoryId:
          type: string
        aisleId:
          type: string
        itemTypeId:
          type: string
        unitType:
          type: string
        isWeighed:
          type: boolean
        isSellable:
          type: boolean
        isStocked:
          type: boolean
        packagingUnitName:
          type: string
        packagingUnitQty:
          type: number
        bundleQty:
          type: integer
          format: int32
        bundlePrice:
          type: number
        bundleName:
          type: string
        minStockLevel:
          type: number
        reorderLevel:
          type: number
        reorderQty:
          type: number
        expiresAfterDays:
          type: integer
          format: int32
        hasExpiry:
          type: boolean
        imageKey:
          type: string
          maxLength: 2048
        active:
          type: boolean
        version:
          type: integer
          format: int64
        images:
          type: array
          items:
            $ref: '#/components/schemas/ItemImageResponse'
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ItemSummaryResponse'
    ItemSummaryResponse:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        barcode:
          type: string
        name:
          type: string
        variantName:
          type: string
        categoryId:
          type: string
        imageKey:
          type: string
          maxLength: 2048
        thumbnailUrl:
          type: string
          description: HTTPS URL for list thumbnails; first gallery image when imageKey is not a delivery URL
          maxLength: 2048
        active:
          type: boolean
        variantOfItemId:
          type: string
...
    CreateVariantRequest:
      type: object
      properties:
        sku:
          type: string
          maxLength: 191
          minLength: 0
        variantName:
          type: string
          maxLength: 255
          minLength: 0
        barcode:
          type: string
          maxLength: 191
          minLength: 0
        name:
          type: string
          maxLength: 500
          minLength: 0
        description:
          type: string
          maxLength: 10000
          minLength: 0
        categoryId:
          type: string
          maxLength: 36
          minLength: 0
        aisleId:
          type: string
          maxLength: 36
          minLength: 0
        unitType:
          type: string
          maxLength: 16
          minLength: 0
        isWeighed:
          type: boolean
        isSellable:
          type: boolean
        isStocked:
          type: boolean
        minStockLevel:
          type: number
        reorderLevel:
          type: number
        reorderQty:
          type: number
        imageKey:
          type: string
          maxLength: 2048
          minLength: 0
      required:
      - sku
      - variantName
    RegisterItemImageRequest:
      type: object
      description: Register already-uploaded image metadata. Provide either legacy s3Key or Cloudinary secureUrl + cloudinaryPublicId.
      properties:
        s3Key:
          type: string
          maxLength: 512
        sortOrder:
          type: integer
          format: int32
          maximum: 9999
          minimum: 0
        width:
          type: integer
          format: int32
          maximum: 32767
          minimum: 1
        height:
          type: integer
          format: int32
          maximum: 32767
          minimum: 1
        contentType:
          type: string
          maxLength: 128
        altText:
          type: string
          maxLength: 500
        primary:
          type: boolean
        secureUrl:
          type: string
          maxLength: 2048
        cloudinaryPublicId:
          type: string
          maxLength: 512
        bytes:
          type: integer
          format: int64
        format:
          type: string
          maxLength: 32
        assetSignature:
          type: string
          maxLength: 80
        predominantColorHex:
          type: string
          maxLength: 16
        phash:
          type: string
          maxLength: 64
...
    LinkedPriceRuleRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        precedence:
          type: integer
          format: int32
      required:
      - id
      - name
      - precedence
    EffectivePricingContextResponse:
      type: object
      properties:
        itemId:
          type: string
        categoryId:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        inheritedMarkupPercent:
          type: number
          nullable: true
        markupSourceKey:
          type: string
        resolvedTaxRateId:
          type: string
          nullable: true
        resolvedTaxRateName:
          type: string
          nullable: true
        resolvedTaxRatePercent:
          type: number
          nullable: true
        taxInclusive:
          type: boolean
          nullable: true
        taxSourceKey:
          type: string
        linkedPriceRules:
          type: array
          items:
            $ref: '#/components/schemas/LinkedPriceRuleRef'
...
    PatchItemRequest:
      type: object
      properties:
        barcode:
          type: string
          maxLength: 191
          minLength: 0
        name:
          type: string
          maxLength: 500
          minLength: 0
        description:
          type: string
          maxLength: 10000
          minLength: 0
        categoryId:
          type: string
          maxLength: 36
          minLength: 0
        aisleId:
          type: string
          maxLength: 36
          minLength: 0
        unitType:
          type: string
          maxLength: 16
          minLength: 0
        isWeighed:
          type: boolean
        isSellable:
          type: boolean
        isStocked:
          type: boolean
        packagingUnitName:
          type: string
          maxLength: 255
          minLength: 0
        packagingUnitQty:
          type: number
        bundleQty:
          type: integer
          format: int32
        bundlePrice:
          type: number
        bundleName:
          type: string
          maxLength: 255
          minLength: 0
        minStockLevel:
          type: number
        reorderLevel:
          type: number
        reorderQty:
          type: number
        expiresAfterDays:
          type: integer
          format: int32
        hasExpiry:
          type: boolean
        imageKey:
          type: string
          maxLength: 2048
          minLength: 0
        active:
          type: boolean
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageItemSummaryResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/ItemSummaryResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
...
  securitySchemes:
    bearer:
      type: http
      description: 'Slice 3: access JWT via Authorization header.'
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      description: 'Slice 3: tenant integration key (prefix + secret; returned once
        at create).'
      name: X-API-Key
      in: header
    cookieRefresh:
      type: apiKey
      description: Optional HttpOnly refresh cookie; native clients may use JSON body
        instead.
      name: ub_refresh
      in: cookie